wed midday update

+++ src/main/java/org/example/Repository/SellerRepository.java
@@ -3,5 +3,5 @@ package org.example.Repository;
 import org.example.Model.Seller;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SellerRepository extends JpaRepository<Seller, Integer> {
+public interface SellerRepository extends JpaRepository<Seller, Long> {
 }
